[mod-sbs] Fix handling of hyphens property
$hyphenation is now a top-level switch for disabling automatic
hyphenation. `hyphens:auto` will only work when hyphenation is true, and
will otherwise behave like `hyphens:manual`.

The expected output in the test was modified because it was
misleading. Some hyphenation points were the result of automatic
hyphenation, not a direct mapping of the manual hyphenation point in the
input.

see issue https://github.com/sbsdev/pipeline-mod-sbs/issues/13