Makefile: automatically update versions in dependencyManagement during release
Because these versions are not taken care of by the
maven-release-plugin, the Makefile previously forced you to update the
versions manually before a release could be made. This meant that a
module may have had a snapshot version while the non-snapshot version
was used elsewhere in a dependency declaration. As a consequence the
Makefile needed to do some guess work which could lead to situations
like described in issue https://github.com/daisy/pipeline/pull/523.

A second benefit of automatically updating versions is that it doesn't
need to be done (manually) in the super repo anymore. This was
problematic because the changes always needed to be braught over to
the subrepo sides which made the release process more manual than
needed.