[modules] Ace validation module
New internal (java based) xproc step and corresponding exposed xproc
step  and library to allow the use of a system wide install of the
Accessibility Checker for EPUB to create an accessibility compliance
report.

The presence of ACE within the path of the user is checked when the
provider is activated.

The report cannot be exported "as-is" due to some non-closed html tags,
but it is reloaded using the html-load step of the html-utils module.
An issue will be open in the ACE github repo to request a correction of
the html or the generation of a valid xhtml report if possible.