[modules] Add new "stylesheet-parameters" option to braille scripts
With this option one can specify values for XSLT parameters and SASS
variables used in style sheets that were provided through the
"stylesheets" option.

The parameters need to be specified as a single string in the query
syntax. This is done because we currently don't have a way to specify
real (XProc) parameters to scripts.

A new px:parse-query step was added to convert a query to a
c:param-set document.

Also fixed a bug in SassCompiler: special characters need to be
escaped according to CSS rules.