Don't echo echo commands in the Makefile
Normally make always echos commands. If the command itself is an echo
command it does not need to be echoed as the sole purpose of it is to
print something.

The special built-in target .SILENT makes sure that for those rules
that are just meant for printing the recipe is not printed also.

This mkes sure that only the text from the echo statements is actually
printed, instead of both the text and the echo commnd itself.