[modules] Add support for "OneCore" voices
On the native side:

- Added a "onecorenative" project to generate the onecore/winrt
  compatible DLL
- Refactored the two native projects
- Also updated the projects to use Visual Studio 2022 and updated tool
  chain

On the Java side: There is still a single TTSEngine for both legacy
SAPI and OneCore voices. getAvailableVoices() returns both kind of
voices, and synthesize() dispatches to the correct underlying
implementation.

Also replaced the workaround for the "unsupported ssml:mark tag" issue
in the SAPI connector (see commit d2cad09) with a better fix: an
xml:lang attribute is now added if missing. (This happens at the
native level.) Now both the SAPI and OneCore connectors use SSML
marks, and ending marks are correctly handled.