[braille-css] SimpleInlineStyle: enable inheriting if and only if parent style is provided
If inheriting is enabled, all inheriting properties defined on the
parent style (but not on the child style) are inherit, and "inherit"
values are always concretized.

Inheriting from a parent style without concretizing "inherit" is not
possible anymore.

Concretizing "inherit" when there is no parent style is expected to be
done only on the root.