[modules] Rewrite native Onecore connector in attempt to fix stack overrun exception
Multithreaded calls to the Onecore synthesizer may trigger a FastFail
error (0xC0000409) on the native side that we are not able to catch
and handle. The following changes were done in an attempt to fix this:

- Global definitions and connections logic was removed on the native
  side to improve thread data isolation.
- Native code now directly returns voice list. Before, the voices were
  constructed in Java from various native method calls that were using
  global variables.
- Native code now directly returns NativeSynthesisResult object
  holding the audio byte array, array of mark names and array of
  positions.

The same improvements were done on the legacy SAPI side (although the
error does not happen with legacy SAPI).

Additional utilities have been added to the jni helper header.